requirements.txt
streamlit
google-generativeai
SpeechRecognition
streamlit-mic-recorder